# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Cloudexus API",
        "version": "1",
        "summary": "The catalog, stock, prices, invoices, partners and orders, and the warehouse app's bookings.",
        "description": "JSON in and out; every message in English. Amounts are in the primary currency (meta.currency); translatable text comes in the language asked for (meta.language). The whole description, for people, is web/API.md — and inside the application under API → API documentation.",
        "license": {
            "name": "AGPL-3.0-only",
            "identifier": "AGPL-3.0-only"
        }
    },
    "servers": [
        {
            "url": "https://cloudexus.example/api"
        }
    ],
    "security": [
        {
            "token": []
        },
        {
            "apiKey": []
        }
    ],
...
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "operationId": "products",
                "summary": "Products",
                "description": "q searches SKU, name and barcode.",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/q"
                    },
                    {
                        "$ref": "#/components/parameters/updated_since"
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "One category."
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        },
                        "description": "Active or inactive only."
                    },
                    {
                        "$ref": "#/components/parameters/language"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A page.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PageMeta"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/Unprocessable"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
...
        "/products/{id}": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "operationId": "product",
                "summary": "A product",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    },
                    {
                        "$ref": "#/components/parameters/language"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The product.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductDetail"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Meta"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
...
        "/warehouses": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "operationId": "warehouses",
                "summary": "Warehouses",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/q"
                    },
                    {
                        "$ref": "#/components/parameters/updated_since"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        },
                        "description": "Active or inactive only."
                    },
                    {
                        "$ref": "#/components/parameters/language"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A page.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Warehouse"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PageMeta"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/Unprocessable"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "scheme": "bearer",
                "description": "An integration token (API → API users) or a user token from POST /auth/login (cxu_…)."
            },
            "apiKey": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Api-Key",
                "description": "The same token, for servers that strip the Authorization header."
            }
        },
        "parameters": {
...
            "page": {
                "name": "page",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 200,
                    "default": 50
                }
            },
            "q": {
                "name": "q",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string"
                },
                "description": "Free-text search; the fields searched differ per list."
            },
            "updated_since": {
                "name": "updated_since",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string",
                    "examples": [
                        "2026-07-19 00:00:00"
                    ]
                },
                "description": "Only records changed since then: for a delta sync."
            },
            "language": {
                "name": "language",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string",
                    "examples": [
                        "en"
                    ]
                },
                "description": "The language translatable text comes in (see GET /languages); a missing translation falls back to the default. An unknown code is 422."
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "Invalid or missing API token.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
            "Unprocessable": {
                "description": "The request does not make sense; details may say which lines.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
        },
        "schemas": {
...
            "PageMeta": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    },
                    "currency": {
                        "$ref": "#/components/schemas/Currency"
                    },
                    "language": {
                        "$ref": "#/components/schemas/Language"
                    }
                },
                "required": [
                    "page",
                    "per_page",
                    "total",
                    "total_pages"
                ]
            },
...
        }
    }
}